requirements/dev-3.9.txt
alabaster==0.7.16
    # via sphinx
asttokens==3.0.0
    # via stack-data
async-timeout==5.0.1
    # via redis
attrs==25.1.0
    # via
    #   jsonschema
    #   referencing
babel==2.17.0
    # via sphinx
black==25.1.0
    # via authedwig
boto3==1.36.23
    # via
    #   authedwig
    #   moto
boto3-stubs[sns,sqs]==1.36.23
    # via authedwig
botocore==1.36.23
    # via
    #   boto3
    #   moto
    #   s3transfer
botocore-stubs==1.36.22
    # via boto3-stubs
build==1.2.2.post1
    # via pip-tools
cachetools==5.5.1
    # via google-auth
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   black
    #   pip-tools
coverage[toml]==7.6.12
    # via
    #   authedwig
    #   coveralls
    #   pytest-cov
coveralls==4.0.1
    # via authedwig
cryptography==44.0.1
    # via moto
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
docopt==0.6.2
    # via coveralls
docutils==0.21.2
    # via sphinx
exceptiongroup==1.2.2
    # via
    #   ipython
    #   pytest
executing==2.2.0
    # via stack-data
factory-boy==3.3.3
    # via authedwig
faker==36.1.1
    # via factory-boy
flake8==7.1.2
    # via authedwig
freezegun==1.5.1
    # via authedwig
funcy==2.0
    # via authedwig
google-api-core[grpc]==2.24.1
    # via google-cloud-pubsub
google-auth==2.38.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
google-cloud-pubsub==2.21.4
    # via authedwig
googleapis-common-protos[grpc]==1.63.1
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.13.0
    # via google-cloud-pubsub
grpcio==1.70.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.48.2
    # via
    #   authedwig
    #   google-api-core
    #   google-cloud-pubsub
idna==3.10
    # via requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==8.5.0
    # via
    #   build
    #   opentelemetry-api
    #   sphinx
iniconfig==2.0.0
    # via pytest
ipdb==0.13.13
    # via authedwig
ipython==8.18.1
    # via ipdb
jedi==0.19.2
    # via ipython
jinja2==3.1.5
    # via
    #   moto
    #   sphinx
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpointer==3.0.0
    # via authedwig
jsonschema==4.23.0
    # via authedwig
jsonschema-specifications==2024.10.1
    # via jsonschema
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib-inline==0.1.7
    # via ipython
mccabe==0.7.0
    # via flake8
moto[sns,sqs]==5.0.28
    # via authedwig
mypy==1.15.0
    # via authedwig
mypy-boto3-sns==1.36.3
    # via boto3-stubs
mypy-boto3-sqs==1.36.0
    # via boto3-stubs
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
opentelemetry-api==1.30.0
    # via
    #   authedwig
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.30.0
    # via authedwig
opentelemetry-semantic-conventions==0.51b0
    # via opentelemetry-sdk
packaging==24.2
    # via
    #   black
    #   build
    #   pytest
    #   sphinx
parso==0.8.4
    # via jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0
    # via ipython
pip-tools==7.4.1
    # via authedwig
platformdirs==4.3.6
    # via black
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.50
    # via ipython
proto-plus==1.26.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
protobuf==3.19.6
    # via
    #   authedwig
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycodestyle==2.12.1
    # via flake8
pycparser==2.22
    # via cffi
pyflakes==3.2.0
    # via flake8
pygments==2.19.1
    # via
    #   ipython
    #   sphinx
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.3.4
    # via
    #   authedwig
    #   pytest-cov
    #   pytest-env
pytest-cov==6.0.0
    # via authedwig
pytest-env==1.1.5
    # via authedwig
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   freezegun
    #   moto
pyyaml==6.0.2
    # via responses
redis==5.2.1
    # via authedwig
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   coveralls
    #   google-api-core
    #   moto
    #   responses
    #   sphinx
responses==0.25.6
    # via moto
retrying==1.3.4
    # via authedwig
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
s3transfer==0.11.2
    # via boto3
six==1.17.0
    # via
    #   python-dateutil
    #   retrying
snowballstemmer==2.2.0
    # via sphinx
sphinx==7.4.7
    # via
    #   authedwig
    #   sphinx-autodoc-typehints
sphinx-autodoc-typehints==2.3.0
    # via authedwig
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
stack-data==0.6.3
    # via ipython
structlog==25.1.0
    # via authedwig
tomli==2.2.1
    # via
    #   black
    #   build
    #   coverage
    #   ipdb
    #   mypy
    #   pip-tools
    #   pytest
    #   pytest-env
    #   sphinx
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
types-awscrt==0.23.10
    # via botocore-stubs
types-dataclasses==0.6.6
    # via authedwig
types-protobuf==3.19.22
    # via authedwig
types-s3transfer==0.11.2
    # via boto3-stubs
typing-extensions==4.12.2
    # via
    #   black
    #   boto3-stubs
    #   ipython
    #   mypy
    #   mypy-boto3-sns
    #   mypy-boto3-sqs
    #   opentelemetry-sdk
    #   referencing
    #   structlog
tzdata==2025.1
    # via faker
urllib3==1.26.20
    # via
    #   botocore
    #   requests
    #   responses
wcwidth==0.2.13
    # via prompt-toolkit
werkzeug==3.1.3
    # via moto
wheel==0.45.1
    # via
    #   authedwig
    #   pip-tools
wrapt==1.17.2
    # via deprecated
xmltodict==0.14.2
    # via moto
zipp==3.21.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
